+++ packages/pangea-node-sdk/CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - AI Guard: `transformed` field.
 
+### Changed
+
+- Retries now include a request header `X-Pangea-Retried-Request-Ids` to track
+  the request IDs of the retries.
+
 ## 5.2.0 - 2025-07-21
 
 ### Added